ref: split lastWillservice in 2
